+++ pipeline/process_block.go
@@ -5,6 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"os"
+	"runtime/debug"
 
 	"connectrpc.com/connect"
 	"github.com/streamingfast/bstream"
@@ -26,6 +28,8 @@ import (
 	"google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
+
+var PrintStack = os.Getenv("SUBSTREAMS_PRINT_STACK") == "true" || os.Getenv("SUBSTREAMS_PRINT_STACK") == "1"
 
 var ErrShuttingDown = errors.New("endpoint is shutting down, please reconnect")
 var minBlocksProcessedToSave = uint64(25)
@@ -322,6 +326,9 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 					err = connect.NewError(connect.CodeDeadlineExceeded, fmt.Errorf("execution timed out at block %d: [%s] %s", blockNum, execOutput.Clock().Id, r))
 				} else {
 					err = fmt.Errorf("panic at block %d: %s [%s]", blockNum, r, execOutput.Clock().Id)
+					if PrintStack {
+						debug.PrintStack()
+					}
 				}
 			}
 		}
@@ -331,6 +338,7 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 	maxParallelExecutor := reqctx.MaxStageLayerParallelExecutor(ctx)
 	logging.Logger(ctx, p.stores.logger).Debug("executing stage's layers", zap.Int("layer_count", len(p.StagedModuleExecutors)), zap.Uint64("max_parallel_executor", maxParallelExecutor))
 
+	executedStages := make(map[int]bool)
 	for _, layer := range p.StagedModuleExecutors {
 		if maxParallelExecutor <= 1 || len(layer) <= 1 {
 			for _, executor := range layer {
@@ -338,6 +346,9 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 					continue
 				}
 				res := p.execute(ctx, executor, execOutput, isFinalBlock)
+				if res.output != nil && !res.skipped_output && !res.output.Cached {
+					executedStages[p.moduleNameToStage[res.output.ModuleName]] = true
+				}
 				if err := p.applyExecutionResult(ctx, executor, res, execOutput); err != nil {
 					return fmt.Errorf("applying executor results %q on block %d (%s): %w", executor.Name(), blockNum, execOutput.Clock().Id, res.err)
 				}
@@ -365,6 +376,9 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 									err = connect.NewError(connect.CodeDeadlineExceeded, fmt.Errorf("execution timed out at block %d: [%s] %s", blockNum, execOutput.Clock().Id, r))
 								} else {
 									err = fmt.Errorf("panic at block %d: %s [%s]", blockNum, r, execOutput.Clock().Id)
+									if PrintStack {
+										debug.PrintStack()
+									}
 								}
 							}
 						}
@@ -385,6 +399,9 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 				if result.not_runnable {
 					continue
 				}
+				if result.output != nil && !result.skipped_output && !result.output.Cached {
+					executedStages[p.moduleNameToStage[result.output.ModuleName]] = true
+				}
 				executor := layer[i]
 				if result.err != nil {
 					return fmt.Errorf("running executor %q: %w", executor.Name(), result.err)
@@ -396,6 +413,7 @@ func (p *Pipeline) executeModules(ctx context.Context, execOutput execout.Execut
 			}
 		}
 	}
+	metering.AddProcessedBlocks(ctx, len(executedStages)) // blocks are counted on every stage for which they were executed (not skipped for indexes, not loaded from existing cache)
 
 	return nil
 }